+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- "Remove" label to option zero in the quantity selector dropdown.
+
 ## [0.11.2] - 2019-10-18
 
 ### Added